update metodos de gasto
